Project1
We are using the sample data collected from the "archive.ics.uci.edu " Website . using this data we are developing an machine learning model to Predict Some has an breast cancer or not. Given data will be analyzed and make it ready for the machine learning algorithm. we are using the three different algorithms to make an pretty accurate predictions of data. the foolowing are the algorithms used they are RandomForestClassifier , KNeighborsClassifier,Support Vector Machine. we will be comparing the scores of the each and three model and save the best model.